Implements jumping to the just created jobs/volumes etc.
